[KYUUBI #4035] Post SparkOperationEvent for all SparkOperations and show sessionId for statements
### _Why are the changes needed?_

Before, we can only see sql query on the kyuubi query engine tab.
<img width="1727" alt="image" src="https://user-images.githubusercontent.com/6757692/209688435-5fc45e5f-2aa9-4991-9cad-1268fc61fc66.png">

After this pr, we can see both scala & python code in the kyuubi query engine tab.

<img width="1727" alt="image" src="https://user-images.githubusercontent.com/6757692/209742783-9867d42d-b04d-403d-872a-c501db483ac6.png">

Also support to show sessionId.
![image](https://user-images.githubusercontent.com/6757692/209756690-b3eb5465-5d29-4983-9178-c86361eb9d13.png)

It also close #2886
### _How was this patch tested?_
- [ ] Add some test cases that check the changes thoroughly including negative and positive cases if possible

- [x] Add screenshots for manual tests if appropriate

- [x] [Run test](https://kyuubi.apache.org/docs/latest/develop_tools/testing.html#running-tests) locally before make a pull request

Closes #4035 from turboFei/scala_python_event.

Closes #4035

33c3a4879 [fwang12] refactor
20dfa874e [fwang12] fix ut
d68fce9a3 [fwang12] support progress
8a86e3235 [fwang12] refactor
568ef8624 [fwang12] sessionId
1d623a29a [fwang12] h link
613e829bb [fwang12] show session id
2f120f1d3 [fwang12] add operation listener
529d91c89 [fwang12] move to spark operation
af21769a7 [fwang12] SQL => Statement
713c593ce [fwang12] scala python

Authored-by: fwang12 <fwang12@ebay.com>
